+++ include/fluxgl/fluxgl.h
@@ -9,5 +9,6 @@
 #include "graphics/vertex.h"
 #include "graphics/shader.h"
 #include "graphics/mesh.h"
+#include "graphics/texture.h"
 #include "graphics/material.h"
 #include "graphics/renderer.h"